+++ geometry/hittable.h
@@ -9,9 +9,9 @@ struct Material;
 struct HitRecord {
     point3 p;
     vec3 normal;
-    double t;
+    float t;
 
-    double u, v;
+    float u, v;
 
     Material const* material;
     bool front_face;
