+++ ray.h
@@ -13,6 +13,6 @@ inline __device__ struct Ray make_ray(point3 origin, vec3 direction)
     return (struct Ray) { .origin = origin, .direction = direction };
 }
 
-inline __device__ point3 ray_at(struct Ray r, double t) { return r.origin + t * r.direction; }
+inline __device__ point3 ray_at(struct Ray r, float t) { return r.origin + t * r.direction; }
 
 #endif
