+++ textures/solid_texture.h
@@ -9,6 +9,6 @@ struct SolidTextureData {
 
 struct SolidTextureData make_solid_texture_data(color c);
 struct Texture make_solid_texture(struct SolidTextureData* d);
-__device__ color solid_texture_color_at(struct SolidTextureData* d, double, double, point3);
+__device__ color solid_texture_color_at(struct SolidTextureData* d, float, float, point3);
 
 #endif
