+++ util/rng.cu
@@ -1,14 +1,14 @@
 #include <math.h>
 #include <util/rng.h>
 
-__device__ double rng_next(curandState_t* state)
+__device__ float rng_next(curandState_t* state)
 {
     // curand_uniform returns between the range (0.0, 1] but we
     // actually prefer the range [0.0, 1.0)
     return 1.0 - curand_uniform(state);
 }
 
-__device__ double rng_next_in_range(curandState_t* state, double min, double max)
+__device__ float rng_next_in_range(curandState_t* state, float min, float max)
 {
     return min + rng_next(state) * (max - min);
 }
@@ -16,18 +16,18 @@ __device__ double rng_next_in_range(curandState_t* state, double min, double max
 __device__ vec3 rng_next_in_unit_sphere(curandState_t* state)
 {
     // See: https://karthikkaranth.me/blog/generating-random-points-in-a-sphere/
-    double u = rng_next(state);
-    double v = rng_next(state);
-    double theta = u * 2.0 * M_PI;
-    double phi = cos(2.0 * v - 1.0);
-    double r = cbrt(rng_next(state));
-    double sin_theta = sin(theta);
-    double cos_theta = cos(theta);
-    double sin_phi = sin(phi);
-    double cos_phi = cos(phi);
-    double x = r * sin_phi * cos_theta;
-    double y = r * sin_phi * sin_theta;
-    double z = r * cos_phi;
+    float u = rng_next(state);
+    float v = rng_next(state);
+    float theta = u * 2.0 * M_PI;
+    float phi = cos(2.0 * v - 1.0);
+    float r = cbrt(rng_next(state));
+    float sin_theta = sin(theta);
+    float cos_theta = cos(theta);
+    float sin_phi = sin(phi);
+    float cos_phi = cos(phi);
+    float x = r * sin_phi * cos_theta;
+    float y = r * sin_phi * sin_theta;
+    float z = r * cos_phi;
 
     return make_vec3(x, y, z);
 }
